Update PortalSpawnHandler.java
adjusts spawning of mirage spots to prevent them in the nether and end, also prevents them spawning over void.

They are specifically prevented in the end, but for other dimension mods, they might work in the other dimensions, but shouldn't spawn over the void anymore.

This addresses https://github.com/Pokecube-Development/Pokecube-Issues-and-Wiki/issues/152